moving code to utils
